Refactor portfolio evidence handling in OpenBadgeCredentialCard and VerifiableCredentialCard.
