+++ app/lib/credentialDisplay/openBadgeCredential.tsx
@@ -26,6 +26,7 @@ import {
   getSubject,
   AlignmentsList
 } from './shared'
+import { portfolioEvidenceFrom } from './shared/utils/evidence'
 
 import { DATE_FORMAT } from '../../../app.config'
 import { getCredentialName } from '../credentialName'
@@ -72,27 +73,7 @@ const OpenBadgeCredentialCard = ({
     alignment
   } = credentialSubjectRenderInfoFrom(subject)
 
-  const portfolioEvidence = (() => {
-    const raw = subject?.portfolio
-    const items = Array.isArray(raw) ? raw : raw ? [raw] : []
-
-    return items
-      .map((item: unknown) => {
-        if (typeof item === 'string') {
-          const url = item.trim()
-          if (!url) return null
-          return { name: url, url }
-        }
-        if (!item || typeof item !== 'object') return null
-
-        const url = String((item as any).url ?? (item as any).id ?? '').trim()
-        if (!url) return null
-
-        const name = String((item as any).name ?? '').trim()
-        return { name: name || url, url }
-      })
-      .filter(Boolean) as Array<{ name: string; url: string }>
-  })()
+  const portfolioEvidence = portfolioEvidenceFrom(subject?.portfolio)
 
   const issuedToName: string = issuedTo || (name as string)
   const credentialName = getCredentialName(credential)
