+++ app/lib/credentialDisplay/verifiableCredential.tsx
@@ -23,6 +23,7 @@ import {
   IssuerInfoButton,
   getSubject
 } from './shared'
+import { portfolioEvidenceFrom } from './shared/utils/evidence'
 
 import { DATE_FORMAT } from '../../../app.config'
 import { getCredentialName } from '../credentialName'
@@ -60,27 +61,7 @@ function VerifiableCredentialCard({
     endDateFmt
   } = credentialSubjectRenderInfoFrom(subject)
 
-  const portfolioEvidence = (() => {
-    const raw = subject?.portfolio
-    const items = Array.isArray(raw) ? raw : raw ? [raw] : []
-
-    return items
-      .map((item: unknown) => {
-        if (typeof item === 'string') {
-          const url = item.trim()
-          if (!url) return null
-          return { name: url, url }
-        }
-        if (!item || typeof item !== 'object') return null
-
-        const url = String((item as any).url ?? (item as any).id ?? '').trim()
-        if (!url) return null
-
-        const name = String((item as any).name ?? '').trim()
-        return { name: name || url, url }
-      })
-      .filter(Boolean) as Array<{ name: string; url: string }>
-  })()
+  const portfolioEvidence = portfolioEvidenceFrom(subject?.portfolio)
 
   const { issuerName, issuerUrl, issuerId, issuerImage } =
     issuerRenderInfoWithVerification(issuer, verifyCredential?.result)
